requirements.txt
pyspark
delta-spark==2.4.0
requests
beautifulsoup4
python-dotenv